Moved Manager\TransactionManager to Transaction\TransactionManager and Manager\UpdateManager and Updater\UpdateManager
