+++ src/DoctrineAuditBundle/Resources/config/services.yaml
@@ -32,12 +32,12 @@ services:
     public: true
 
   dh_doctrine_audit.manager:
-    class: DH\DoctrineAuditBundle\Manager\TransactionManager
+    class: DH\DoctrineAuditBundle\Transaction\TransactionManager
     arguments:
       - "@dh_doctrine_audit.configuration"
     public: true
 
-  DH\DoctrineAuditBundle\Manager\TransactionManager:
+  DH\DoctrineAuditBundle\Transaction\TransactionManager:
     alias: dh_doctrine_audit.manager
     public: true
